Label the Docker workspace volume mount shared
Container processess spawned by the Docker workflow plugin are using the
default SELinux context, which does not allow for filesystem access in
locations such as `/var`, for instance. Therefore, volumes mounted into
the container should be re-labeled with to be shared so that the
container processes can share access to the mounted data.

See also:
http://www.projectatomic.io/blog/2015/06/using-volumes-with-docker-can-cause-problems-with-selinux/

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>