Removing defined stop timeout
When a SIGTERM is sent the process is safely pruned however the existing limit of 1 second is not always enough for the application to respect the signal. As if this is breached and causes the pipeline to fail this can cause any larger containers to be unusable. Dockers default value is 10 seconds, extending to 5 seconds, ideally this could be customizable if others have similar concerns