Fix host directory mounts on Windows
On Windows, it is not possible to mount host directories to a subfolder
of a non-C: drive. Due to this restriction, having a workspace on
another drive will make the docker run execution fail since the plugin
tries to map the workspace into the container.

This fixes that by collecting all the non-C: drive volume mappings and
handling those explicitly by mapping the entire drive into the
container.