Run containers with --init by default to avoid leaking zombie processes.
People running containers manually usually run with pid 1 being a
$SHELL, in which case waiting for child termination is something
implemented already as it's the core job of a $SHELL. However with this
jenkins plugin the command run is usually "cat" on Linux, which doesn't
wait at all for children. To have in jenkins a situation a bit like
during interactive container experience, start containers with --init
so zombie processes are correctly reaped.