Use `docker ps` to detect absence of docker permissions
The plugin BOM agents seem to have the `docker` command available but the
user running the agent is not authorized to use the `docker` command.
Previously that was detected by calling `docker ps` and detecting the
failure.

Restores a change made in

* https://github.com/jenkinsci/docker-workflow-plugin/pull/331

Testing done

Confirmed that I could see the same failure on a local computer as is seen on
https://ci.jenkins.io/job/Tools/job/bom/job/master/3968/testReport/org.jenkinsci.plugins.docker.workflow/DockerDSLTest/

The computer had Docker CE installed by the specific user running the
test did not have permission to access Docker.  Prior to this change,
the tests failed with the message:

CANNOT CONNECT TO THE DOCKER DAEMON AT UNIX:///VAR/RUN/DOCKER.SOCK. IS THE DOCKER DAEMON RUNNING?

After making this change, the tests pass on that computer with the
specific user that does not have permission to access Docker.