Pattern adapted to determine whether Jenkins is running in a container when podman is used with cgroup v2
The docker-workflow plugin also works perfectly with podman. After switching podman to cgroup v2, it is no longer recognised when Jenkins is running in a podman container. In order for this to work again, the regular expression needs to be adjusted. After the adjustment, the containerid is recognised correctly for both docker and podman. The expression used utilises lookbehind and lookahead expressions to determine the containerid.