Fix WithContainerStep when hasWorkdir is false for Windows Containers
The existing logic in the Decorator class sets hasWorkDir if the docker
version is newer than 17.12. In the case where hasWorkDir is false, the
fallback logic in the decorate() method updates the starter cmds() but
always uses a UNIX-like "sh" command, which causes Jenkins to hang and
timeout when using Windows Containers.

Update the logic in the decorate() method to provide an implementation
for Windows Containers that prepends the appropriate "cd" command
followed by && to the generated command.